сделал отдельные модули под каждый тестируемый модуль
