+++ pyproject.toml
@@ -23,6 +23,10 @@ black = "^25.1.0"
 isort = "^6.0.1"
 
 
+
+[tool.poetry.group.dev.dependencies]
+pytest-cov = "^6.2.1"
+
 [tool.isort]
 line_length = 119
